api/pyproject.toml
[project]
name = "assetflow-api"
version = "0.1.0"
description = "AssetFlow FastAPI backend"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
  "alembic>=1.14.0",
  "celery[redis]>=5.4.0",
  "email-validator>=2.2.0",
  "eval-type-backport>=0.2.0",
  "fastapi>=0.115.0",
  "httpx>=0.28.0",
  "passlib[bcrypt]>=1.7.4",
  "psycopg[binary]>=3.2.0",
  "pydantic-settings>=2.6.0",
  "pyjwt>=2.10.0",
  "python-multipart>=0.0.20",
  "sqlalchemy>=2.0.36",
  "uvicorn[standard]>=0.34.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.3.0",
  "pytest-asyncio>=0.24.0",
  "pytest-cov>=6.0.0",
  "ruff>=0.9.0",
]

[build-system]
requires = ["hatchling>=1.26.0"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]

[tool.pytest.ini_options]
pythonpath = ["."]
testpaths = ["tests"]

[tool.ruff]
line-length = 100
target-version = "py39"

[tool.ruff.lint]
select = ["E", "F", "B", "N", "S", "A"]
ignore = ["B008", "E501", "S101", "S105"]
